Made color palettes, memory map, various cpu values, and scanline buffer 32-bit aligned for speed (another 6% boost with my test frame!) The majority of my test ROMs now play at 100% speed with 0 frameskip when overclocked :-)

+++ src/display.h
@@ -1,7 +1,7 @@
 #pragma once
 
 void SetupDisplayDriver(bool withStretch, char withFrameskip);
-void SetupDisplayPalette(unsigned short pal[12]);
+void SetupDisplayPalette(unsigned int pal[12]);
 
 void DmaWaitNext(void);
